# docs/03_API仕様/openapi.yaml
openapi: 3.1.0
info:
  title: 人事考課システム API
  description: |
    トキワテック人事考課システムのREST API仕様書。

    ## 認証
    Cloudflare Access (JWT) による認証を使用します。
    全てのAPIリクエストには `Authorization: Bearer <token>` ヘッダーが必要です。

    ## 共通仕様
    - 日時はISO 8601形式 (例: `2026-01-24T10:00:00Z`)
    - IDはUUID v4形式
    - ページネーションはoffset/limit方式
  version: 1.0.0
  contact:
    name: トキワテック開発チーム
    email: dev@tokiwa-tech.com
  license:
    name: Proprietary
    identifier: LicenseRef-proprietary
servers:
  - url: https://evaluation.tokiwa-tech.com/api/v1
    description: 本番環境
  - url: https://evaluation-staging.tokiwa-tech.com/api/v1
    description: ステージング環境
  - url: http://localhost:8787/api/v1
    description: ローカル開発環境
...
paths:
...
  /cycles:
    get:
      tags: [Cycles]
      summary: 自分の評価サイクル一覧
      operationId: listMyCycles
      security:
        - BearerAuth: []
      parameters:
        - name: periodId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 評価サイクル一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CycleListResponse'
  /cycles/{cycleId}:
    get:
      tags: [Cycles]
      summary: 評価サイクル詳細取得
      operationId: getCycle
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      responses:
        '200':
          description: 評価サイクル詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CycleDetailResponse'
  /admin/cycles:
    get:
      tags: [Admin, Cycles]
      summary: 全評価サイクル一覧
      description: 管理者のみ
      operationId: listAllCycles
      security:
        - BearerAuth: []
      parameters:
        - name: periodId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CycleStatus'
        - name: departmentId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 全評価サイクル一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CycleListResponse'

  # ============================================
  # 自己評価 API (3 endpoints)
  # ============================================
  /cycles/{cycleId}/self-evaluation:
    get:
      tags: [Self Evaluation]
      summary: 自己評価取得
      operationId: getSelfEvaluation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      responses:
        '200':
          description: 自己評価データ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfEvaluationResponse'

    put:
      tags: [Self Evaluation]
      summary: 自己評価保存（下書き）
      operationId: saveSelfEvaluation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSelfEvaluationRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfEvaluationResponse'
  /cycles/{cycleId}/self-evaluation/submit:
    post:
      tags: [Self Evaluation]
      summary: 自己評価提出
      operationId: submitSelfEvaluation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  description: 提出コメント
      responses:
        '200':
          description: 提出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================
  # 同僚評価 API (5 endpoints)
  # ============================================
  /cycles/{cycleId}/peer-evaluations:
    get:
      tags: [Peer Evaluation]
      summary: 同僚評価一覧取得
      operationId: listPeerEvaluations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      responses:
        '200':
          description: 同僚評価一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerEvaluationListResponse'

  # 注: 同僚評価の詳細操作は /cycles/:cycleId/peer-evaluations/:peerId/* に統合

  # ============================================
  # 上司評価 API (6 endpoints)
  # ============================================
  /manager/cycles:
    get:
      tags: [Manager Evaluation]
      summary: 部下の評価サイクル一覧
      operationId: listSubordinateCycles
      security:
        - BearerAuth: []
      parameters:
        - name: periodId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CycleStatus'
      responses:
        '200':
          description: 部下の評価サイクル一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CycleListResponse'
  /cycles/{cycleId}/manager-evaluation:
    get:
      tags: [Manager Evaluation]
      summary: 上司評価取得
      operationId: getManagerEvaluation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      responses:
        '200':
          description: 上司評価データ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerEvaluationResponse'

    put:
      tags: [Manager Evaluation]
      summary: 上司評価保存
      operationId: saveManagerEvaluation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveManagerEvaluationRequest'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerEvaluationResponse'
  /cycles/{cycleId}/manager-evaluation/submit:
    post:
      tags: [Manager Evaluation]
      summary: 上司評価提出
      operationId: submitManagerEvaluation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      responses:
        '200':
          description: 提出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /cycles/{cycleId}/approve:
    post:
      tags: [Manager Evaluation, HR]
      summary: 評価承認
      operationId: approveCycle
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        '200':
          description: 承認成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /cycles/{cycleId}/reject:
    post:
      tags: [Manager Evaluation, HR]
      summary: 評価差戻し
      operationId: rejectCycle
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CycleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
      responses:
        '200':
          description: 差戻し成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================
  # HR API (統合済み - cycles/:cycleId/approve, reject で対応)
  # ============================================
  # 注: HR承認機能は /cycles/:cycleId/approve, /cycles/:cycleId/reject に統合
  # requireRole('HR', 'ADMIN') でアクセス制御

  # ============================================
  # 社員 API (5 endpoints)
  # ============================================
  /employees:
    get:
      tags: [Employees]
      summary: 社員一覧
      operationId: listEmployees
      security:
        - BearerAuth: []
      parameters:
        - name: departmentId
          in: query
          schema:
            type: string
        - name: gradeId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 社員一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
  /employees/{employeeId}:
    get:
      tags: [Employees]
      summary: 社員詳細取得
      operationId: getEmployee
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '200':
          description: 社員詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailResponse'

    put:
      tags: [Employees]
      summary: 社員情報更新
      operationId: updateEmployee
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetailResponse'
...
  /notifications:
    get:
      tags: [Notifications]
      summary: 通知一覧
      operationId: listNotifications
      security:
        - BearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 通知一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
  /notifications/{notificationId}/read:
    put:
      tags: [Notifications]
      summary: 既読にする
      operationId: markNotificationRead
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 既読化成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: 全て既読にする
      operationId: markAllNotificationsRead
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 全既読化成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================
  # ダッシュボード API (2 endpoints)
  # ============================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cloudflare Access JWT token
  parameters:
...
    CycleId:
      name: cycleId
      in: path
      required: true
      schema:
        type: string
      description: 評価サイクルID
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
      description: 社員ID
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: ページ番号
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 1ページあたりの件数
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean

    # ============================================
    # 列挙型
    # ============================================
    CycleStatus:
      type: string
      enum:
        - DRAFT
        - SELF_SUBMITTED
        - PEER_COMPLETED
        - MANAGER_SUBMITTED
        - MANAGER_APPROVED
        - HR_APPROVED
        - FINALIZED
        - REJECTED
    AxisCode:
      type: string
      enum: [SKILL, COMPETENCY, BEHAVIOR]
    Rank:
      type: string
      enum: [S, A, B, C, D]
...
    Period:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        evaluationMonth:
          type: integer
          minimum: 1
          maximum: 12
        status:
          type: string
          enum: [PREPARING, ACTIVE, CLOSED]
...
    Cycle:
      type: object
      properties:
        id:
          type: string
        periodId:
          type: string
        employeeId:
          type: string
        status:
          $ref: '#/components/schemas/CycleStatus'
        currentStep:
          type: string
        finalScore:
          type: integer
          nullable: true
        finalRank:
          $ref: '#/components/schemas/Rank'
    CycleListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Cycle'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CycleDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/Cycle'
            - type: object
              properties:
                period:
                  $ref: '#/components/schemas/Period'
                employee:
                  $ref: '#/components/schemas/EmployeeSummary'
                selfEvaluation:
                  $ref: '#/components/schemas/SelfEvaluationSummary'
                peerEvaluations:
                  type: array
                  items:
                    $ref: '#/components/schemas/PeerEvaluationSummary'
                managerEvaluation:
                  $ref: '#/components/schemas/ManagerEvaluationSummary'
                approvalHistory:
                  type: array
                  items:
                    $ref: '#/components/schemas/ApprovalHistory'
    EmployeeSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        department:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
    SelfEvaluationSummary:
      type: object
      properties:
        id:
          type: string
        submittedAt:
          type: string
          format: date-time
          nullable: true
    PeerEvaluationSummary:
      type: object
      properties:
        id:
          type: string
        evaluator:
          $ref: '#/components/schemas/EmployeeSummary'
        submittedAt:
          type: string
          format: date-time
          nullable: true
    ManagerEvaluationSummary:
      type: object
      nullable: true
      properties:
        id:
          type: string
        submittedAt:
          type: string
          format: date-time
          nullable: true
        totalScore:
          type: number
          nullable: true
    ApprovalHistory:
      type: object
      properties:
        step:
          type: string
        action:
          type: string
        actorId:
          type: string
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ============================================
    # 自己評価
    # ============================================
    SelfEvaluationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            cycleId:
              type: string
            comment:
              type: string
              nullable: true
            scores:
              type: array
              items:
                $ref: '#/components/schemas/EvaluationScore'
            submittedAt:
              type: string
              format: date-time
              nullable: true
    EvaluationScore:
      type: object
      properties:
        roadmapItemId:
          type: string
        roadmapItem:
          $ref: '#/components/schemas/RoadmapItemSummary'
        score:
          type: integer
          minimum: 1
          maximum: 5
        evidence:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
    RoadmapItemSummary:
      type: object
      properties:
        itemNumber:
          type: integer
        category:
          $ref: '#/components/schemas/AxisCode'
        title:
          type: string
    SaveSelfEvaluationRequest:
      type: object
      properties:
        comment:
          type: string
        scores:
          type: array
          items:
            type: object
            properties:
              roadmapItemId:
                type: string
              score:
                type: integer
                minimum: 1
                maximum: 5
              evidence:
                type: string
              comment:
                type: string

    # ============================================
    # 同僚評価
    # ============================================
    PeerEvaluationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/PeerEvaluation'
    PeerEvaluation:
      type: object
      properties:
        id:
          type: string
        cycle:
          type: object
          properties:
            id:
              type: string
            period:
              $ref: '#/components/schemas/Period'
        targetEmployee:
          $ref: '#/components/schemas/EmployeeSummary'
        submittedAt:
          type: string
          format: date-time
          nullable: true
...
    ManagerEvaluationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            cycleId:
              type: string
            scores:
              type: array
              items:
                $ref: '#/components/schemas/EvaluationScore'
            skillScore:
              type: number
              nullable: true
            competencyScore:
              type: number
              nullable: true
            behaviorScore:
              type: number
              nullable: true
            totalScore:
              type: number
              nullable: true
            rankSuggestion:
              $ref: '#/components/schemas/Rank'
            overallComment:
              type: string
              nullable: true
            submittedAt:
              type: string
              format: date-time
              nullable: true
    SaveManagerEvaluationRequest:
      type: object
      properties:
        scores:
          type: array
          items:
            type: object
            properties:
              roadmapItemId:
                type: string
              score:
                type: integer
                minimum: 1
                maximum: 5
              comment:
                type: string
        skillScore:
          type: number
        competencyScore:
          type: number
        behaviorScore:
          type: number
        totalScore:
          type: number
        rankSuggestion:
          $ref: '#/components/schemas/Rank'
        overallComment:
          type: string
    RejectRequest:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
        rejectToStep:
          type: string
          enum: [SELF, PEER, MANAGER]
...
    Employee:
      type: object
      properties:
        id:
          type: string
        employeeCode:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          $ref: '#/components/schemas/Department'
        grade:
          $ref: '#/components/schemas/Grade'
        currentStep:
          type: integer
        manager:
          $ref: '#/components/schemas/EmployeeSummary'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, LEAVE]
        joinedAt:
          type: string
          format: date
    EmployeeListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        pagination:
          $ref: '#/components/schemas/Pagination'
    EmployeeDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Employee'
    UpdateEmployeeRequest:
      type: object
      properties:
        departmentId:
          type: string
        gradeId:
          type: string
        managerId:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, LEAVE]

    # ============================================
    # スキル達成状況
    # ============================================
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        parentId:
          type: string
          nullable: true
...
    Grade:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        level:
          type: integer
        description:
          type: string
          nullable: true
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        link:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================
    # ダッシュボード
    # ============================================
...